moved utility functions so they don't show up as endpoint
